--- packages/core/examples/image_stack_from_omezarr3d_f32/index.html
@@ -1,76 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Z-Stack Viewer</title>
-    <style>
-      html,
-      body {
-        margin: 0;
-        padding: 0;
-        display: flex;
-        width: 100%;
-        height: 100%;
-      }
-      #main {
-        display: flex;
-        flex-direction: column;
-        width: 100%;
-        height: 100%;
-      }
-      #canvas {
-        flex: 1;
-        min-height: 0;
-      }
-      #controls {
-        margin: 0.5rem 2rem;
-        font-family: sans-serif;
-      }
-      .control-row {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        margin-bottom: 1rem;
-      }
-      .control-label {
-        width: 120px;
-      }
-      .slider-container {
-        flex: 1;
-        margin: 0 1rem;
-      }
-      .value-label {
-        width: 120px;
-        text-align: right;
-      }
-      input[type="range"] {
-        width: 100%;
-      }
-      input[type="button"] {
-        margin-left: 1rem;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="main">
-      <canvas id="canvas"></canvas>
-      <div id="controls">
-        <div class="control-row">
-          <div class="control-label">Z Slice:</div>
-          <div class="slider-container">
-            <input type="range" id="z-slider" />
-          </div>
-          <div class="value-label">
-            <span id="z-index">0</span> / <span id="z-total">0</span>
-          </div>
-        </div>
-        <div class="control-row">
-          Layer State: &nbsp; <span id="layer-state">initialized</span>
-          <input type="button" id="load-all" value="Load all slices" />
-        </div>
-      </div>
-    </div>
-    <script type="module" src="./main.ts"></script>
-  </body>
-</html>
