+++ packages/core/examples/index.html
@@ -35,11 +35,6 @@
         >
       </li>
       <li>
-        <a href="/image_stack_from_omezarr3d_f32/index.html"
-          >2D Image Series from 3D OMEZarr with f32 data (lazy loading)</a
-        >
-      </li>
-      <li>
         <a href="/image2d_from_omezarr5d_u8/index.html"
           >2D Image from 5D OMEZarr with uint8 data</a
         >
